requirements.txt
# 基础依赖
requests>=2.31.0
python-dotenv>=1.0.0
loguru>=0.7.2

# 数据库 (SQLite - Python内置，无需额外安装)
sqlalchemy>=2.0.23

# AI和NLP
openai>=1.3.0
langchain>=0.0.340
tiktoken>=0.5.2

# 数据处理
pandas>=2.1.4
numpy>=1.24.0
openpyxl>=3.1.0  # Excel导出支持

# 日志和配置
pydantic>=2.5.0
